+++ tests/test_solution.py
@@ -35,8 +35,14 @@ def s4():
 
 @pytest.fixture()
 def s5():
-    # 100 mg/L as CaCO3
+    # 100 mg/L as CaCO3 ~ 1 mM
     return Solution([["Ca+2", "40.078 mg/L"], ["CO3-2", "60.0089 mg/L"]], volume="1 L")
+
+
+@pytest.fixture()
+def s5_pH():
+    # 100 mg/L as CaCO3 ~ 1 mM
+    return Solution([["Ca+2", "40.078 mg/L"], ["CO3-2", "60.0089 mg/L"]], volume="1 L", balance_charge="pH")
 
 
 @pytest.fixture()
@@ -55,6 +61,26 @@ def s6():
             ["Br-", "20 mM"],
         ],  # -20 meq/L
         volume="1 L",
+    )
+
+
+@pytest.fixture()
+def s6_Ca():
+    # non-electroneutral solution with lots of hardness
+    # alk = -118 meq/L * 50 = -5900 mg/L, hardness = 12*50 = 600 mg/L as CaCO3
+    # charge balance = 2+10+10+10-120-20-12 = -120 meq/L
+    return Solution(
+        [
+            ["Ca+2", "1 mM"],  # 2 meq/L
+            ["Mg+2", "5 mM"],  # 10 meq/L
+            ["Na+1", "10 mM"],  # 10 meq/L
+            ["Ag+1", "10 mM"],  # no contribution to alk or hardness
+            ["CO3-2", "6 mM"],  # no contribution to alk or hardness
+            ["SO4-2", "60 mM"],  # -120 meq/L
+            ["Br-", "20 mM"],
+        ],  # -20 meq/L
+        volume="1 L",
+        balance_charge="Ca+2",
     )
 
 
@@ -153,18 +179,23 @@ def test_chempot_energy(s1, s2):
     pass
 
 
-def test_alkalinity_hardness_chargebalance(s3, s5, s6):
+def test_charge_balance(s3, s5, s5_pH, s6, s6_Ca):
     assert np.isclose(s3.charge_balance, 0)
+    assert np.isclose(s5.charge_balance, 0, atol=1e-5)
+    assert np.isclose(s5_pH.charge_balance, 0, atol=1e-8)
+    assert np.isclose(s6.charge_balance, -0.12)
+    assert np.isclose(s6_Ca.charge_balance, 0, atol=1e-8)
+
+
+def test_alkalinity_hardness(s3, s5, s6):
     assert np.isclose(s3.hardness, 0)
     assert np.isclose(s3.alkalinity, 0)
 
     assert np.isclose(s5.alkalinity.magnitude, 100, rtol=0.005)
     assert np.isclose(s5.hardness.magnitude, 100, rtol=0.005)
-    assert np.isclose(s5.charge_balance, 0, atol=1e-5)
 
     assert np.isclose(s6.alkalinity.magnitude, -5900, rtol=0.005)
     assert np.isclose(s6.hardness.magnitude, 600, rtol=0.005)
-    assert np.isclose(s6.charge_balance, -0.12)
 
 
 def test_pressure_temperature(s5):